backend/requirements.txt
fastapi
uvicorn
google-auth
earthengine-api
